requirements.txt
# Core LLM + Transformers
torch
transformers
accelerate
sentencepiece
tiktoken

# Streamlit + UI
streamlit
streamlit_autorefresh

# Metrics + System Stats
psutil
GPUtil

# Utilities
numpy
pandas
pyyaml
tqdm

# Plotting / Visualizations
matplotlib
seaborn

# (Optional but recommended)
scikit-learn
altair
pandas
streamlit